Reorganize Wear Material list composables
- Move list-related composables to dedicated package
- Remove redundant `OssLicenseListImpl` layer
- Split components into focused files
